requirements.txt
# Numina backend - matches pyproject.toml [project].dependencies
# Requires Python >=3.10

# Web
fastapi>=0.128.0,<0.129.0
uvicorn>=0.40.0,<0.41.0

# Config & validation
pydantic-settings>=2.12.0,<3.0.0
email-validator>=2.0.0,<3.0.0

# Database
sqlalchemy>=2.0.46,<3.0.0
asyncpg>=0.31.0,<0.32.0
alembic>=1.18.2,<2.0.0

# Auth
pyjwt>=2.9.0,<3.0.0
passlib[bcrypt]>=1.7.4,<2.0.0

# Redis & workers
redis>=7.1.0,<8.0.0
arq>=0.26.0,<0.28.0

# Optional integrations
stripe>=14.3.0,<15.0.0
openai>=1.0.0,<2.0.0
httpx>=0.27.0,<0.28.0
